+++ plugins/lookup/app/javascript/reverse_app/test/testHelper.js
@@ -1,0 +1,7 @@
+import ReduxThunk from 'redux-thunk'
+import configureStore from 'redux-mock-store'
+
+export function setupStore(initialState) {
+  const mockStore = configureStore([ReduxThunk])
+  return mockStore(initialState)
+}
